Search Folder Created, .env Initialized
Components about search are divided to several files and moved to "search" folder.
