+++ Application/Mobile/babel.config.js
@@ -2,5 +2,13 @@ module.exports = {
   presets: ['module:metro-react-native-babel-preset'],
   plugins: [
     'react-native-reanimated/plugin',
-  ],
+    ["module:react-native-dotenv", {
+      "envName": "APP_ENV",
+      "moduleName": "@env",
+      "path": ".env",
+      "safe": false,
+      "allowUndefined": true,
+      "verbose": false
+    }]
+  ]
 };
