+++ Application/Mobile/package.json
@@ -34,6 +34,7 @@
     "eslint": "^8.26.0",
     "jest": "^29.2.2",
     "metro-react-native-babel-preset": "^0.73.3",
+    "react-native-dotenv": "^3.4.2",
     "react-test-renderer": "^18.1.0"
   },
   "jest": {
