+++ Application/Mobile/pages/home/HomeHeader.js
@@ -1,17 +1,21 @@
 import React from 'react';
-import { Appbar, Searchbar } from 'react-native-paper';
-import { View, StyleSheet } from 'react-native'
+import { Appbar } from 'react-native-paper';
+import { View, StyleSheet, Image } from 'react-native'
 
 // The top navigation bar of home feed
-const HomeHeader = ({navigation}) => {
+const HomeHeader = ({ navigation }) => {
 
     return (
         <Appbar.Header style={styles.topBar}>
             <Appbar.Action icon="menu" color="white" onPress={() => navigation.toggleDrawer()} />
-        <View style={styles.rightContent}>
-            <Appbar.Action icon="magnify" color='white' onPress={() => navigation.navigate('Search')} />
-            <Appbar.Action icon="robot-happy" color='white' onPress={() => navigation.navigate('Chatbot')} />
-        </View>
+            <View style={styles.logoContainer}>
+                <Appbar.Action style={styles.logo} color ='white' icon={require('../../images/logo.png')}
+                />
+            </View>
+            <View style={styles.rightContent}>
+                <Appbar.Action icon="magnify" color='white' onPress={() => navigation.navigate('Search')} />
+                <Appbar.Action icon="robot-happy" color='white' onPress={() => navigation.navigate('Chatbot')} />
+            </View>
         </Appbar.Header>
     );
 }
@@ -24,7 +28,7 @@ const styles = StyleSheet.create({
         justifyContent: 'space-between',
         alignItems: 'center',
     },
-    searchBar:  {
+    searchBar: {
         width: '65%',
         height: '70%',
         borderWidth: 1,
@@ -37,5 +41,8 @@ const styles = StyleSheet.create({
     },
     searchInput: {
         fontSize: 13,
+    },
+    logoContainer: {
+        position: 'absolute', top: 0, left: 0, right: 0, bottom: 0, justifyContent: 'center', alignItems: 'center'
     },
 });
