+++ Application/Mobile/pages/search/SearchScreenHeader.js
