+++ Application/Mobile/pages/userAPI.js
@@ -1,6 +1,6 @@
 import { AsyncStorage } from 'react-native';
 import { handleLogoutRequest } from './authAPI';
-
+import {BACKEND_URL} from '@env'
 
 export const handleGetUserData = async () => {
     const token = await AsyncStorage.getItem("token")
@@ -13,7 +13,7 @@ export const handleGetUserData = async () => {
         redirect: 'follow'
     };
 
-    return await fetch("http://18.206.229.240:8000/profile/", requestOptions)
+    return await fetch(BACKEND_URL + "profile/", requestOptions)
         .then(function (response) {
             return response.json();
         })
